requirements.txt
altair==5.1.2
anyio==4.0.0
appnope==0.1.3
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
arrow==1.3.0
astroid==3.0.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
executing==2.0.1
fastjsonschema==2.18.1
fonttools==4.43.1
fqdn==1.5.1
gitdb==4.0.11
gitpython==3.1.40
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython-genutils==0.2.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
jedi==0.19.1
jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema-specifications==2023.7.1
jsonschema==4.19.2
jsonschema[format-nongpl]==4.19.2
jupyter-client==8.5.0
jupyter-console==6.6.3
jupyter-core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter-server-terminals==0.4.4
jupyter-server==2.9.1
jupyter==1.0.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.0
jupyterlab-widgets==3.0.9
jupyterlab==4.0.8
kiwisolver==1.4.5
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib-inline==0.1.6
matplotlib==3.8.1
mccabe==0.7.0
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.2
mypy-extensions==1.0.0
mypy==1.6.1
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook-shim==0.2.3
notebook==7.0.6
numpy==1.26.1
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandas-stubs==2.1.1.230928
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycparser==2.21
pydeck==0.8.0
pygments==2.16.1
pylint==3.0.2
pyparsing==3.1.1
pysocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
pyyaml==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
qtpy==2.4.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
selenium==4.15.2
send2trash==1.8.2
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.28.2
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tomlkit==0.12.1
toolz==0.12.0
tornado==6.3.3
traitlets==5.13.0
trio-websocket==0.11.1
trio==0.23.1
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ucimlrepo==0.0.3
uri-template==1.3.0
urllib3==2.0.7
urllib3[socks]==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
wsproto==1.2.0
zipp==3.17.0
